Lahmian sires become queen's direct children
From an RP point-of-view, vampire sires becoming the vampire's children
makes a lot of sense, and that's how it works in this mod for the most
part already.

For the Lahmians however, the recruitment mission is the main way to
grow your dynasty. But since the queen can't go on that mission herself,
the new members are always grand-children, or grand-grand-children etc.
From a CK2 perspective, with title inheritance and such, it makes a huge
mess.

I suggest vampires recruited through that mission end up being the
queen's children instead.

Also break any existing marriage so they don't just fly back to their
old courts. It's much better with your vampires buddies isn't it?